vcspull.repo.BaseRepo.run(): Keep track of the stdout output buffer
This commits update run() function so that it sets the a new attribute
named "stdout_data" to process object returned by the function.

Additionally, it adds the parameter "log_stdout" allowing to disable
the logging of the stdout. This is particularly useful when the
caller want to explicitly use the stdout buffer without have it
displayed in the terminal.