Rf get_url_and_rev to get_url_and_rev_from_pip_url
The long method name is temporary.  Currently, the url attribute
on VCS classes is the  pip-style url and not the raw URL a repo
would use.

This leads to confusion and ambiguity since the pip-url is really
only important from a front-end perspective. After the true url
and rev is resolved, there is no need for it in the instance.